+++ examples/javascript/instance-principal.js
@@ -17,7 +17,7 @@ const common = require("oci-common");
 (async () => {
   const authenticationProvider = await new common.InstancePrincipalsAuthenticationDetailsProviderBuilder().build();
   try {
-    const identityClient = await new identity.IdentityClient({
+    const identityClient = new identity.IdentityClient({
       authenticationDetailsProvider: authenticationProvider
     });
     const regions = await identityClient.listRegions({});