+++ lib/managementdashboard/lib/model/management-dashboard-for-import-export-details.ts
@@ -101,6 +101,10 @@ export interface ManagementDashboardForImportExportDetails {
    */
   "parametersConfig"?: Array<any>;
   /**
+   * Drill-down configuration to define the destination of a drill-down action.
+   */
+  "drilldownConfig"?: Array<any>;
+  /**
    * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
    * Example: `{\"bar-key\": \"value\"}`
    *