+++ lib/managementdashboard/lib/model/update-management-dashboard-details.ts
@@ -93,6 +93,10 @@ export interface UpdateManagementDashboardDetails {
    */
   "parametersConfig"?: Array<any>;
   /**
+   * Drill-down configuration to define the destination of a drill-down action.
+   */
+  "drilldownConfig"?: Array<any>;
+  /**
    * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
    * Example: `{\"bar-key\": \"value\"}`
    *