+++ lib/operatoraccesscontrol/lib/model/operator-control-summary.ts
@@ -39,6 +39,10 @@ export interface OperatorControlSummary {
    */
   "isFullyPreApproved"?: boolean;
   /**
+   * resourceType for which the OperatorControl is applicable
+   */
+  "resourceType"?: model.ResourceTypes;
+  /**
    * Time when the operator control was created, expressed in [RFC 3339] (https://tools.ietf.org/html/rfc3339) timestamp format. Example: '2020-05-22T21:10:29.600Z'
    *
    */