+++ lib/operatoraccesscontrol/lib/model/resource-types.ts
@@ -24,6 +24,8 @@ import common = require("oci-common");
  **/
 export enum ResourceTypes {
   Exacc = "EXACC",
+  Exadatainfrastructure = "EXADATAINFRASTRUCTURE",
+  Autonomousvmcluster = "AUTONOMOUSVMCLUSTER",
 
   /**
    * This value is used if a service returns a value for this enum that is not recognized by this